+++ src/felice/serialization.clj
@@ -1,10 +1,13 @@
 (ns ^:no-doc felice.serialization
   (:require [cognitect.transit :as transit]
-            [jsonista.core :as json])
+            [jsonista.core :as json]
+            [taoensso.nipp :as nippy])
   (:import [org.apache.kafka.common.serialization Serializer       Deserializer
                                                   LongSerializer   LongDeserializer
                                                   StringSerializer StringDeserializer]
            [java.io ByteArrayInputStream ByteArrayOutputStream]))
+
+;;; transit
 
 (defn transit-serializer [type]
   (reify
@@ -31,6 +34,8 @@
   (json/object-mapper {:encode-key-fn name
                        :decode-key-fn keyword
                        :date-format "yyyy-MM-dd'T'HH:mm:ss.SSSX"}))
+
+;;; json
 
 (defn json-serializer []
   (reify
@@ -66,19 +71,45 @@
                {:raw-value as-string
                 ::error {:deserializing e}}))))))
 
+;;; nippy
+
+
+(defn nippy-serializer ^Serializer []
+  (reify
+    Serializer
+    (close [this])
+    (configure [this config is-key?])
+    (serialize [this topic payload]
+      (nippy/fast-freeze payload))))
+
+(defn nippy-deserializer ^Deserializer []
+  (reify
+    Deserializer
+    (close [this])
+    (configure [this config is-key?])
+    (deserialize [this topic payload]
+      (try
+        (nippy/fast-thaw payload)
+        (catch Exception e
+          (throw (ex-info "corrupted nippy byte array"
+                          {:cause e
+                           :topic topic})))))))
+
 
 (def serializers {:long     (fn [] (LongSerializer.))
                   :string   (fn [] (StringSerializer.))
                   :json     json-serializer
                   :t+json   (partial transit-serializer :json)
-                  :t+mpack  (partial transit-serializer :msgpack)})
+                  :t+mpack  (partial transit-serializer :msgpack)
+                  :nippy    nippy-serializer})
 
 (def deserializers {:long    (fn [] (LongDeserializer.))
                     :string  (fn [] (StringDeserializer.))
                     :json    json-deserializer
                     :json-safe    json-safe-deserializer
                     :t+json  (partial transit-deserializer :json)
-                    :t+mpack (partial transit-deserializer :msgpack)})
+                    :t+mpack (partial transit-deserializer :msgpack)
+                    :nippy  nippy-deserializer})
 
 (defn ^Serializer serializer [s]
   (if (keyword? s)